+++ src/viewOpenGL/MiniMapGL.cpp
@@ -7,10 +7,10 @@
 
 static const QColor minimap_view_color = Qt::blue;
 static const QColor minimap_scene_color = Qt::red;
-static const qreal minimap_height = 100.0;
-static const qreal minimap_width = 100.0;
+static const float minimap_height = 100.0;
+static const float minimap_width = 100.0;
 
-MiniMapGL::MiniMapGL(QObject* parent)
+MiniMapGL::MiniMapGL(QObject *parent)
     : GraphicItemGL(parent)
     , m_sceneColor(minimap_scene_color)
     , m_viewColor(minimap_view_color)
@@ -34,21 +34,21 @@ void MiniMapGL::clearData()
     m_scene = QRectF();
 }
 
-void MiniMapGL::setScene(const QRectF& scene)
+void MiniMapGL::setScene(const QRectF &scene)
 {
     if (scene.isValid() && scene != m_scene) {
         m_scene = scene;
     }
 }
 
-void MiniMapGL::setViewPort(const QRectF& view)
+void MiniMapGL::setViewPort(const QRectF &view)
 {
     if (view.isValid() && m_viewPort != view) {
         m_viewPort = view;
     }
 }
 
-void MiniMapGL::setParentSceneTransformations(const QTransform& transform)
+void MiniMapGL::setParentSceneTransformations(const QTransform &transform)
 {
     if (m_parentSceneTransformations != transform) {
         m_parentSceneTransformations = transform;
@@ -63,7 +63,7 @@ QTransform MiniMapGL::localTransform() const
 
     const QSizeF maxBoundingSize(minimap_height, minimap_width);
     const QSizeF sceneScaledSize = m_scene.size().scaled(maxBoundingSize, Qt::KeepAspectRatio);
-    const qreal scaleFactor = sceneScaledSize.height() / m_scene.height();
+    const float scaleFactor = sceneScaledSize.height() / static_cast<float>(m_scene.height());
 
     QTransform transform = QTransform::fromScale(scaleFactor, scaleFactor);
     const QPointF top_left = transform.mapRect(m_scene).topLeft();
@@ -71,7 +71,7 @@ QTransform MiniMapGL::localTransform() const
     return transform;
 }
 
-void MiniMapGL::doDraw(QOpenGLFunctionsVersion& opengl_functions)
+void MiniMapGL::doDraw(QOpenGLFunctionsVersion &opengl_functions)
 {
     if (!m_viewPort.isValid() || !m_scene.isValid()) {
         return;
@@ -85,7 +85,7 @@ void MiniMapGL::doDraw(QOpenGLFunctionsVersion& opengl_functions)
                    opengl_functions);
 }
 
-void MiniMapGL::setSceneColor(const QColor& sceneColor)
+void MiniMapGL::setSceneColor(const QColor &sceneColor)
 {
     if (m_sceneColor != sceneColor) {
         m_sceneColor = sceneColor;
@@ -98,7 +98,7 @@ const QColor MiniMapGL::sceneColor() const
     return m_sceneColor;
 }
 
-void MiniMapGL::setViewColor(const QColor& viewColor)
+void MiniMapGL::setViewColor(const QColor &viewColor)
 {
     if (m_viewColor != viewColor) {
         m_viewColor = viewColor;
@@ -116,11 +116,11 @@ const QRectF MiniMapGL::boundingRect() const
     return localTransform().mapRect(m_scene);
 }
 
-void MiniMapGL::setSelectionArea(const SelectionEvent*)
+void MiniMapGL::setSelectionArea(const SelectionEvent *)
 {
 }
 
-void MiniMapGL::mouseMoveEvent(QMouseEvent* event)
+void MiniMapGL::mouseMoveEvent(QMouseEvent *event)
 {
     // move
     if (event->buttons() & Qt::LeftButton) {
@@ -128,7 +128,7 @@ void MiniMapGL::mouseMoveEvent(QMouseEvent* event)
     }
 }
 
-void MiniMapGL::mousePressEvent(QMouseEvent* event)
+void MiniMapGL::mousePressEvent(QMouseEvent *event)
 {
     // center if left button is pressed down
     if (event->button() == Qt::LeftButton) {
@@ -137,7 +137,7 @@ void MiniMapGL::mousePressEvent(QMouseEvent* event)
     }
 }
 
-void MiniMapGL::mouseReleaseEvent(QMouseEvent* event)
+void MiniMapGL::mouseReleaseEvent(QMouseEvent *event)
 {
     // set selecting to false if released
     if (event->button() == Qt::LeftButton) {
@@ -145,7 +145,7 @@ void MiniMapGL::mouseReleaseEvent(QMouseEvent* event)
     }
 }
 
-void MiniMapGL::centerOnLocalPos(const QPointF& localPoint)
+void MiniMapGL::centerOnLocalPos(const QPointF &localPoint)
 {
     // localPoint
     const QPointF scenePoint
